+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Add Steam Tinker Launch.
 - Change GE Proton version from GE-Proton7-37 to GE-Proton7-55.
 - Change GE Wine version from GE-Proton7-31 to GE-Proton7-43.
+- Change the screenshot program from Shutter to Spectacle.
 
 ## 3.2.1 - The SteamOS 3.4 Update